refactor: SelectContainer 컴포넌트 내의 비즈니스 로직 훅으로 분리 #97
